Specify a default for number of puma threads
Previously if you did not specify the `GEMSTASH_PUMA_THREADS`
environment variable, the value of `:puma_threads` would default
to 0.

This was because `ENV['GEMSTASH_PUMA_THREADS']` results in an empty
string, which returns 0 when `#to_i` is called on it.

I've chosen 16 as the default as it is what has been used as the
default by gemstash.